feat: op code
